merge mistake and refactoring
there was a mistake in the merge (players initial positions weren't edited), fixed.

refactored player class to remove JLabel property from it:

in JGame i made an array list named playersLabels contains the labels of the players. the interface between the player and the labels list can be through "getPlayerLabel" function.