payRent and BuyCities
too many details added related to these two functions as well as a text Area to show the details of each player.

Please see the game and tell me of the bugs found in order to fix them.
PS. I worked on the code of Kord ( where there was a bug on the "non-testing" mode ) , He was changing them while i was working on this version. I don't know whether he finished it or not.
Thank you.