needs work, please read comments
THIS IS NOT A FINAL VERSION
function parameters and passing needs a lot of work, i'll refactor it later.

in playerClass
added groups array list, contains the group numbers that the player have and wrote a function to update it once a player complete a group.

added number of companies and number of railroads integers->used in calculate overallrent function

refactored Monopoly controller